requirements.txt
asgiref==3.11.1
black==26.3.1
click==8.3.1
Django==6.0.3
django-cors-headers==4.7.0
djangorestframework==3.17.1
flake8==7.3.0
isort==8.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.2
pytokens==0.4.1
sqlparse==0.5.5
qrcode==8.2
pillow==12.1.1

